LayoutTests/fast/canvas/canvas-currentPath-crash-expected.txt

Series of tests to ensure that currentPath does not take non-Path values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imgdata[0] is 0
PASS imgdata[1] is 128
PASS imgdata[2] is 0
PASS successfullyParsed is true

TEST COMPLETE
